+++ Examples/SwiftNWSDemo/SwiftNWSDemo/ConditionsModel.swift
@@ -83,6 +83,8 @@ final class ConditionsModel {
     case .invalidHeadlineIdentifier: "The headline identifier is invalid."
     case .invalidLink(let link): "The API linked outside itself: \(link)"
     case .invalidOfficeIdentifier: "The office identifier is invalid."
+    case .invalidProductCode: "The product code is invalid."
+    case .invalidProductLocation: "The product location is invalid."
     case .invalidRedirect: "The weather service returned an invalid redirect."
     case .invalidStationIdentifier: "The weather service returned an invalid station identifier."
     case .invalidZoneIdentifier: "The weather service returned an invalid zone identifier."